Restore the old "print" command, for now
Changing the "print" command to use db_ctf_pprint_cmd has broken all
documented uses, and "examine" does not allow printing variables, or
individual registers.

For now, mpi@ agrees with me that it is "okay" to leave the "print"
command at least intact, and in accordance with the documentation, but
we still want "pp[rint]" to replace the "print" command when it works
better.

ok mpi@?